refactor(type-checker): extract FieldConstraints partial class [slice 2/6]
Moves 7 field-constraint validation methods from PreceptTypeChecker.cs
into PreceptTypeChecker.FieldConstraints.cs.

No behavior changes. 1752 Precept.Tests pass.
